+++ .docs/CHANGELOG.md
@@ -1,6 +1,12 @@
 # CHANGELOG
 
 ## [Unreleased] - 1/28/2024
+- Remove net addr in wallet when not delegating
+
+## [Unreleased] - 1/28/2024
+- Add max amount form button
+
+## [Unreleased] - 1/28/2024
 - Fix UI transactions table Issue
 
 ## [Unreleased] - 1/24/2024